chore: removing stories that Histoire can't compile
