+++ client/utilities/rfc-converters-utils.ts
@@ -1,12 +1,13 @@
 import { range } from 'lodash-es'
 import { DateTime } from 'luxon'
-import { z } from 'zod'
+import type { z } from 'zod'
 import { parseRFCId } from './rfc'
 import type { RfcCommon, RFCJSON } from './rfc'
 import { NONBREAKING_SPACE } from './strings'
 import { assertIsString, assertNever } from './typescript'
 import type { HintedString } from './typescript'
 import type { Rfc, RfcMetadata } from '~/generated/red-client'
+import type { TypeSenseSearchItemSchema } from './typesense'
 
 type RfcMetadataAuthor = RfcMetadata['authors'][number]
 
@@ -235,83 +236,6 @@ export const parseRfcStatusSlug = (
   )
 }
 
-// Schema definition https://github.com/ietf-tools/search/blob/main/schemas/docs.md
-export const TypeSenseSearchItemSchema = z.object({
-  id: z.string(),
-
-  rfcNumber: z.number(),
-  date: z.number(),
-  publicationDate: z.number(),
-
-  title: z.string(),
-
-  stdlevelname: z
-    .enum([
-      'Internet Standard',
-      'Proposed Standard',
-      'Draft Standard',
-      'Best Current Practice',
-      'Informational',
-      'Experimental',
-      'Historic',
-      'Unknown'
-    ])
-    .optional(),
-  abstract: z.string(),
-
-  adName: z.string().optional(),
-  authors: z
-    .array(
-      z.object({
-        name: z.string(),
-        affiliation: z.string()
-      })
-    )
-    .optional(),
-
-  subserieTotal: z.number().optional(),
-  bcp: z.string().optional(),
-  fyi: z.string().optional(),
-  std: z.string().optional(),
-  his: z.string().optional(),
-  rfc: z.string(),
-
-  area: z
-    .object({
-      acronym: z.string(),
-      name: z.string(),
-      full: z.string()
-    })
-    .optional(),
-  group: z.object({
-    acronym: z.string(),
-    name: z.string(),
-    full: z.string()
-  }),
-
-  stream: z
-    .object({
-      slug: z.string(),
-      name: z.string()
-    })
-    .optional(),
-  ranking: z.number(),
-  state: z.array(z.string()),
-
-  type: z.string(),
-
-  filename: z.string(),
-  pages: z.number(),
-  keywords: z.array(z.string()),
-
-  flags: z
-    .object({
-      obsoleted: z.boolean(),
-      updated: z.boolean()
-    })
-    .optional()
-})
-
 export const parseTypeSenseSubseries = (
   item: z.infer<typeof TypeSenseSearchItemSchema>
 ): RfcCommon['subseries'] => {
@@ -325,6 +249,7 @@ export const parseTypeSenseSubseries = (
       }
     case 'Informational':
       return undefined
+    // TODO: use 'Informational' subseries number when available
     // assertIsString(item.fyi)
     // return {
     //   type: 'fyi',
